requirements.txt
# Core requirements
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Web crawling and parsing
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF parsing
pdfplumber>=0.9.0
PyMuPDF>=1.22.0

# Excel parsing
openpyxl>=3.1.0
xlrd>=2.0.0

# Testing
pytest>=7.3.0

# Utilities
python-dotenv>=1.0.0
